Finally got button input and navigation working
I had to debug both the PyGLet and Glooey libraries in order to solve
the problem. It turns out that Glooey is heavily dependent on a lot of
PyGLet methods in ways that are not at all documented. Overriding some
methods for use with direct PyGLet breaks some Glooey functionality, so
I had to prevent this from happening. What I find interesting is that I
have not been able to determine why an override of some of the functions
that Glooey doesn't seem to depend on in particular still breaks things.
This is especially true for the mouse input handling. Perhaps only one
method can process a dispatch of a single mouse event? I don't yet know
enough about these libraries to tell.